Update TestWideFetch to use wide query iter
