+++ src/dbnode/integration/wide_query_test.go
@@ -22,7 +22,6 @@
 
 package integration
 
-/*
 import (
 	"bytes"
 	"fmt"
@@ -128,20 +127,6 @@ func buildExpectedChecksumsByShard(
 		checksums = append(checksums, sharded.checksums...)
 	}
 
-	// NB: IDs should only be included for documents that conclude a batch.
-	l := len(checksums)
-	if l == 0 {
-		return checksums
-	}
-
-	// NB: only look at the last `l-1` elements, as the last element should
-	// always have its ID.
-	for i, checksum := range checksums[:l-1] {
-		if (i+1)%batchSize != 0 {
-			checksums[i].ID = checksum.ID[:0]
-		}
-	}
-
 	return checksums
 }
 
@@ -150,19 +135,20 @@ func assertTags(
 	encodedTags checked.Bytes,
 	decoder serialize.TagDecoder,
 	expected int64,
+	msgAndArgs ...interface{},
 ) {
 	encodedTags.IncRef()
 	encoded := encodedTags.Bytes()
 	encodedTags.DecRef()
 
 	decoder.Reset(checked.NewBytes(encoded, nil))
-	assert.Equal(t, 1, decoder.Len())
-	assert.True(t, decoder.Next())
+	assert.Equal(t, 1, decoder.Len(), msgAndArgs...)
+	assert.True(t, decoder.Next(), msgAndArgs...)
 	tag := decoder.Current()
-	assert.Equal(t, wideTagName, tag.Name.String())
-	assert.Equal(t, fmt.Sprintf(wideTagValFmt, expected), tag.Value.String())
-	assert.False(t, decoder.Next())
-	require.NoError(t, decoder.Err())
+	assert.Equal(t, wideTagName, tag.Name.String(), msgAndArgs...)
+	assert.Equal(t, fmt.Sprintf(wideTagValFmt, expected), tag.Value.String(), msgAndArgs...)
+	assert.False(t, decoder.Next(), msgAndArgs...)
+	require.NoError(t, decoder.Err(), msgAndArgs...)
 }
 
 func assertData(
@@ -185,6 +171,55 @@ func assertData(
 	assert.Equal(t, float64(ex), ts.Value)
 	assert.False(t, it.Next())
 	assert.NoError(t, it.Err())
+}
+
+type testWideQuerySeries struct {
+	Shard            uint32
+	ID               []byte
+	EncodedTags      []byte
+	MetadataChecksum int64
+}
+
+func copyBytes(b []byte) []byte {
+	empty := make([]byte, 0, len(b))
+	return append(empty, b...)
+}
+
+func wideQueryIterSeries(
+	t *testing.T,
+	iter storage.WideQueryIterator,
+) []testWideQuerySeries {
+	var results []testWideQuerySeries
+
+	for iter.Next() {
+		shardIter := iter.Current()
+
+		for shardIter.Next() {
+			seriesIter := shardIter.Current()
+
+			for seriesIter.Next() {
+				// Should get datapoints here when implemented.
+			}
+
+			results = append(results, testWideQuerySeries{
+				Shard:            shardIter.Shard(),
+				ID:               copyBytes(seriesIter.ID().Bytes()),
+				EncodedTags:      copyBytes([]byte(seriesIter.EncodedTags())),
+				MetadataChecksum: seriesIter.MetadataChecksum(),
+			})
+
+			require.NoError(t, seriesIter.Err(), "wide series iter error")
+			seriesIter.Close()
+		}
+
+		require.NoError(t, shardIter.Err(), "wide shard iter error")
+		shardIter.Close()
+	}
+
+	require.NoError(t, iter.Err(), "wide iter error")
+	iter.Close()
+
+	return results
 }
 
 func TestWideFetch(t *testing.T) {
@@ -323,18 +358,22 @@ func TestWideFetch(t *testing.T) {
 	for _, tt := range shardFilterTests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.NewContext()
-			chk, err := testSetup.DB().WideQuery(ctx, nsMetadata.ID(), query,
+			iter, err := testSetup.DB().WideQuery(ctx, nsMetadata.ID(), query,
 				now, tt.shards, iterOpts)
 			require.NoError(t, err)
+
+			chk := wideQueryIterSeries(t, iter)
 
 			expected := buildExpectedChecksumsByShard(ids, tt.shards,
 				testSetup.ShardSet(), batchSize)
 			require.Equal(t, len(expected), len(chk))
 			for i, checksum := range chk {
-				assert.Equal(t, expected[i].MetadataChecksum, checksum.MetadataChecksum)
-				require.Equal(t, string(expected[i].ID), checksum.ID.String())
-				assertTags(t, checksum.EncodedTags, decoder, checksum.MetadataChecksum)
-				checksum.Finalize()
+				assert.Equal(t, expected[i].MetadataChecksum, checksum.MetadataChecksum,
+					fmt.Sprintf("fail checksum match: i=%d", i))
+				assert.Equal(t, expected[i].ID, checksum.ID,
+					fmt.Sprintf("fail id match: i=%d", i))
+				assertTags(t, checksum.EncodedTags, decoder, checksum.MetadataChecksum,
+					fmt.Sprintf("fail metadata checksum match: i=%d", i))
 			}
 
 			ctx.Close()
@@ -391,9 +430,11 @@ func TestWideFetch(t *testing.T) {
 	for _, tt := range exactShardFilterTests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.NewContext()
-			chk, err := testSetup.DB().WideQuery(ctx, nsMetadata.ID(), exactQuery,
+			iter, err := testSetup.DB().WideQuery(ctx, nsMetadata.ID(), exactQuery,
 				now, tt.shards, iterOpts)
 			require.NoError(t, err)
+
+			chk := wideQueryIterSeries(t, iter)
 
 			if !tt.expected {
 				assert.Equal(t, 0, len(chk))
@@ -403,7 +444,6 @@ func TestWideFetch(t *testing.T) {
 				assert.Equal(t, int64(1), checksum.MetadataChecksum)
 				assert.Equal(t, exactID, checksum.ID.String())
 				assertTags(t, checksum.EncodedTags, decoder, checksum.MetadataChecksum)
-				checksum.Finalize()
 			}
 
 			ctx.Close()
@@ -424,13 +464,15 @@ func TestWideFetch(t *testing.T) {
 			var runError error
 			for j := 0; j < runs; j++ {
 				ctx := context.NewContext()
-				chk, err := testSetup.DB().WideQuery(ctx, nsMetadata.ID(), q,
+				iter, err := testSetup.DB().WideQuery(ctx, nsMetadata.ID(), q,
 					now, nil, iterOpts)
 
 				if err != nil {
 					runError = fmt.Errorf("query err: %v", err)
 					break
 				}
+
+				chk := wideQueryIterSeries(t, iter)
 
 				if len(expected) != len(chk) {
 					runError = fmt.Errorf("expected %d results, got %d",
@@ -444,8 +486,6 @@ func TestWideFetch(t *testing.T) {
 							expected[i].MetadataChecksum, checksum.MetadataChecksum)
 						break
 					}
-
-					checksum.Finalize()
 				}
 
 				ctx.Close()
@@ -464,4 +504,3 @@ func TestWideFetch(t *testing.T) {
 	wg.Wait()
 	require.NoError(t, multiErr.LastError())
 }
-*/
